Add makefiles and the sqlite3 amalgamation.
- Add and update makefiles in sim_db/, lib/, test/ and example/.
- The makefiles compile static C and C++ libraries for sim_db called
libsimdbc.a and libsimdbcpp.a.
- Add sqlite3.c and sqlite3.h in third_party/ which is the sqlite3
amalgamation, the entire C-version of sqlite3 put into a single C file.
- Update tests and examples.
- All tests passes.

+++ src/sim_db.cpp
@@ -1,7 +1,7 @@
 // Copyright (C) 2018 Håkon Austlid Taskén <hakon.tasken@gmail.com>
 // Licensed under the MIT License.
 
-#include "sim_db.hpp"
+#include "../include/sim_db.hpp"
 
 namespace sim_db {
 Connection::Connection(int argc, char** argv) {
@@ -28,13 +28,13 @@ std::string Connection::make_unique_subdir(std::string path_directory,
 void Connection::update_sha1_executables(
         std::vector<std::string> paths_executables) {
     char** string_vec = new char*[paths_executables.size()];
-    for (int i = 0; i < paths_executables.size(); i++) {
+    for (size_t i = 0; i < paths_executables.size(); i++) {
         string_vec[i] = new char[paths_executables[i].size() + 1];
         strcpy(string_vec[i], paths_executables[i].c_str());
     }
     sim_db_update_sha1_executables(sim_db, string_vec,
                                    paths_executables.size());
-    for (int i = 0; i < paths_executables.size(); i++) {
+    for (size_t i = 0; i < paths_executables.size(); i++) {
         delete[] string_vec[i];
     }
     delete[] string_vec;
